pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lean4-autoformalization"
version = "0.1.0"
description = "A system for translating natural language mathematical statements into formal Lean 4 code using RAG"
readme = "README.md"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Mathematics"
]
requires-python = ">=3.9"
dependencies = [
    "langchain>=0.1.0",
    "langchain_core>=0.1.0",
    "langchain_community>=0.1.0",
    "langgraph>=0.0.15",
    "pydantic>=2.0.0",
    "chromadb>=0.4.18",
    "langchain_chroma>=0.0.1",
    "langchain_huggingface>=0.0.1",
    "sentence-transformers>=2.2.2",
    "beautifulsoup4>=4.12.2",
    "python-dotenv>=1.0.0",
    "tqdm>=4.66.1"
]

[project.optional-dependencies]
web = ["gradio>=4.0.0"]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0"
]
all = ["lean4-autoformalization[web,dev]"]

[project.urls]
"Homepage" = "https://github.com/yourusername/lean4-autoformalization"
"Bug Tracker" = "https://github.com/yourusername/lean4-autoformalization/issues"

[tool.setuptools]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ["py39"]

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[[tool.mypy.overrides]]
module = [
    "langchain.*",
    "chromadb.*",
    "tqdm.*",
    "gradio.*"
]
ignore_missing_imports = true